Move manage-clients.component and associated API invocation to AC
