+++ libs/common/src/billing/abstractions/billing-api.service.abstraction.ts
@@ -1,5 +1,4 @@
 import { OrganizationCreateRequest } from "../../admin-console/models/request/organization-create.request";
-import { ProviderOrganizationOrganizationDetailsResponse } from "../../admin-console/models/response/provider/provider-organization.response";
 import { SubscriptionCancellationRequest } from "../../billing/models/request/subscription-cancellation.request";
 import { OrganizationBillingMetadataResponse } from "../../billing/models/response/organization-billing-metadata.response";
 import { PlanResponse } from "../../billing/models/response/plan.response";
@@ -29,10 +28,6 @@ export abstract class BillingApiServiceAbstraction {
   abstract getPlans(): Promise<ListResponse<PlanResponse>>;
 
   abstract getProviderClientInvoiceReport(providerId: string, invoiceId: string): Promise<string>;
-
-  abstract getProviderClientOrganizations(
-    providerId: string,
-  ): Promise<ListResponse<ProviderOrganizationOrganizationDetailsResponse>>;
 
   abstract getProviderInvoices(providerId: string): Promise<InvoicesResponse>;
